refactor(AdminController): 전체 예약 대기 조회 API 관리자 권한 부여
